refactor send eth
