+++ examples/rebalancer/src/index.ts
@@ -1,10 +1,9 @@
 import * as path from "path";
 import * as dotenv from "dotenv";
-import { ethers } from "ethers";
-import { toReadableAmount, isKeyOfObject } from "./utils";
+import { isKeyOfObject } from "./utils";
 import { getOrganization, createPrivateKey, createPrivateKeyTag, createUser, createUserTag, createPolicy } from "./requests";
-import { TurnkeyApi, init as httpInit } from "@turnkey/http";
 import { getProvider, getTurnkeySigner } from "./provider";
+import { sendEth } from "./send";
 
 // Load environment variables from `.env.local`
 dotenv.config({ path: path.resolve(process.cwd(), ".env.local") });
@@ -96,7 +95,6 @@ async function fund(args: string[]) {
     const connectedSigner = getTurnkeySigner(provider, bankPrivateKey.privateKeyId);
 
     for (const sourcePrivateKey of sourcePrivateKeys) {
-        // get address
         const ethAddress = sourcePrivateKey.addresses.find(address => {
             return address.format == 'ADDRESS_FORMAT_ETHEREUM';
         });
@@ -104,66 +102,7 @@ async function fund(args: string[]) {
             throw new Error(`couldn't lookup ETH address for private key: ${sourcePrivateKey.privateKeyId}`)
         }
 
-        // transfer eth
-        await transferEth(provider, connectedSigner, ethAddress.address, 1);
+        await sendEth(provider, connectedSigner, ethAddress.address, 1);
     }
 }
 
-async function transferEth(
-    provider: ethers.providers.Provider,
-    connectedSigner: ethers.Signer,
-    destinationAddress: string,
-    value: number,
-) {
-  // TODO(tim): investigate why we can't call `connectedSigner.getNetwork()`
-  const network = await provider.getNetwork();
-  const chainId = await connectedSigner.getChainId();
-  const balance = await connectedSigner.getBalance();
-  const address = await connectedSigner.getAddress();
-
-  print("Network:", `${network.name} (chain ID ${chainId})`);
-  print("Address:", address);
-  print("Balance:", `${ethers.utils.formatEther(balance)} Ether`);
-
-  if (balance.isZero()) {
-      let warningMessage = "The transaction won't be broadcasted because your account balance is zero.\n";
-      if (network.name === "goerli") {
-        warningMessage += "Use https://goerlifaucet.com/ to request funds on Goerli, then run the script again.\n";
-      }
-
-      throw new Error(warningMessage);
-  }
-
-  const feeData = await connectedSigner.getFeeData();
-  const gasRequired = feeData.maxFeePerGas!.mul(21000);
-
-  if (balance.lt(gasRequired.add(value))) {
-    throw new Error(`Insufficient ETH balance. Skipping...`);
-  }
-
-  const transactionRequest = {
-    to: destinationAddress,
-    value,
-    type: 2,
-    maxFeePerGas: feeData.maxFeePerGas!,
-    maxPriorityFeePerGas: feeData.maxPriorityFeePerGas!,
-  };
-
-  const sentTx = await connectedSigner.sendTransaction(transactionRequest);
-  console.log("Awaiting confirmation...");
-
-  await connectedSigner.provider?.waitForTransaction(sentTx.hash, 1);
-
-  print(
-      `Sent ${toReadableAmount(
-        value.toString(),
-        18,
-        12
-      )} ETH to ${destinationAddress}:`,
-      `https://${network.name}.etherscan.io/tx/${sentTx.hash}`
-  );
-}
-
-function print(header: string, body: string): void {
-  console.log(`${header}\n\t${body}\n`);
-}
